backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
openai>=1.12.0
qdrant-client>=1.10.0
python-dotenv>=1.0.0
requests>=2.32.0
pydantic>=2.5.0
python-multipart>=0.0.9
pytest>=7.4.0
httpx>=0.25.0
cohere>=5.10.0
beautifulsoup4>=4.14.0
uv>=0.1.0